+++ back/app/Http/Controllers/Api/StructureController.php
@@ -41,6 +41,9 @@ class StructureController extends Controller
                 AllowedFilter::custom('lieu', new FiltersStructureLieu),
                 AllowedFilter::custom('search', new FiltersStructureSearch),
             ])
+            ->allowedIncludes([
+                'missions', 
+            ])
             ->defaultSort('-updated_at')
             ->paginate($request->input('pagination') ?? config('query-builder.results_per_page'));
     }
@@ -98,7 +101,6 @@ class StructureController extends Controller
     {
         $structure = Structure::with(['members', 'territoire'])->withCount('missions', 'participations', 'waitingParticipations', 'conversations')->where('id', $structure->id)->first();
         $structure->append('response_time_score');
-        ray('structure', $structure);
         return $structure;
     }
 
@@ -165,12 +167,29 @@ class StructureController extends Controller
 
     public function delete(StructureDeleteRequest $request, Structure $structure)
     {
+        if ($structure->missions()->exists())
+        {
+            return response()->json(['errors'=> [
+                'password' => [
+                    "L'organisation ne peut pas être supprimée car elle a des missions liées.",
+                ]
+            ]], 400);
+        }
+        
         return (string) $structure->delete();
+    }
+
+    public function restore($id)
+    {
+        $structure = Structure::withTrashed()->findOrFail($id);
+        $this->authorize('restore', $structure);
+        return (string) $structure->restore();
     }
 
     public function destroy($id)
     {
         $structure = Structure::withTrashed()->findOrFail($id);
+        $this->authorize('destroy', $structure);
         return (string) $structure->forceDelete();
     }
